Create util, constants, algorithms files
Mov
