+++ SokobanGame.py
@@ -1,19 +1,11 @@
+from constants import *
 import tkinter as tk
 import time
-import itertools
-import random
-import copy
+import util
 
 TILESIZE = 40
 XOFFSET = 25
 YOFFSET = 25
-SUPERMALUS = -100
-MALUS = -1
-STEPREWARD = 0
-BONUS = 1
-SUPERBONUS = 100
-MAXSTEPS = 100000
-EPSILONDECAY = 0.995
 
 class Sokoban:
     def __init__(self, master, level_file):
@@ -26,7 +18,6 @@ class Sokoban:
         self.draw_game()
         self.bind_keys()
         self.game_over = False
-        self.action_space = ['up', 'left', 'down', 'right']
 
     def load_level(self):
         with open(self.level_file, "r") as file:
@@ -35,7 +26,7 @@ class Sokoban:
         self.initial_level = [row.copy() for row in self.level]
         self.height = len(self.level)
         self.width = max(len(line) for line in self.level)
-        self.player_pos = self.find_player_in_state(self.level)
+        self.player_pos = util.find_player_in_state(self.level)
         self.total_boxes = sum(row.count("$") for row in self.level)
 
     def create_widgets(self):
@@ -52,7 +43,7 @@ class Sokoban:
 
     def reset_level(self):
         self.level = [row.copy() for row in self.initial_level]
-        self.player_pos = self.find_player_in_state(self.level)
+        self.player_pos = util.find_player_in_state(self.level)
         self.game_over = False
         self.draw_game()
 
@@ -97,7 +88,7 @@ class Sokoban:
             self.master.bind("<Key>", self.key_pressed_human)
 
     def key_pressed_policy(self, event):
-        pass  # Ignore key presses when using a policy
+        pass 
 
     def key_pressed_human(self, event):
         if self.game_over:
@@ -107,9 +98,9 @@ class Sokoban:
             self.reset_level()
             return
 
-        action = self.get_action(event.keysym)
+        action = util.get_action(event.keysym)
         if action:
-            self.player_pos, reward, moved_box = self.execute_action(self.level, action)
+            self.player_pos, reward, moved_box = util.execute_action(self.level, action)
             self.draw_game()
             
             if reward == SUPERBONUS:
@@ -119,321 +110,22 @@ class Sokoban:
                 print("Game lost!")
                 self.game_over = True
 
-    def find_player_in_state(self, state):
-        """
-        Finds the player's position in a given state.
-        Returns (x, y) or None if not found.
-        """
-        for y, row in enumerate(state):
-            for x, cell in enumerate(row):
-                if cell in ('@', '+'):
-                    return (x, y)
-        return None
-
-    def print_state(self, state):
-        for row in state:
-            print(row)
-
-    def generate_state_space(self):
-        print("Generating state space...")
-        state_space = set()
-        wall_positions = set()
-        goal_positions = set()
-        box_positions = set()
-
-        for y, row in enumerate(self.initial_level):
-            for x, cell in enumerate(row):
-                if cell == "#":
-                    wall_positions.add((y, x))
-                elif cell == ".":
-                    goal_positions.add((y, x))
-                elif cell == "$":
-                    box_positions.add((y, x))
-                elif cell == "@":
-                    player_position = (y, x)
-
-        non_wall_positions = [
-            (y, x)
-            for y, row in enumerate(self.initial_level)
-            for x, cell in enumerate(row)
-            if cell != "#"
-        ]
-        # Generate all possible combinations of box positions
-        box_combinations = itertools.combinations(
-            non_wall_positions, len(box_positions)
-        )
-
-        for box_comb in box_combinations:
-            for player_pos in non_wall_positions:
-                if player_pos not in box_comb:
-                    level_state = []
-                    for y, row in enumerate(self.initial_level):
-                        new_row = []
-                        for x, cell in enumerate(row):
-                            pos = (y, x)
-                            if pos in wall_positions:
-                                new_row.append("#")
-                            elif pos in goal_positions and pos in box_comb:
-                                new_row.append("x")
-                            elif pos in goal_positions and pos == player_pos:
-                                new_row.append("+")
-                            elif pos in goal_positions:
-                                new_row.append(".")
-                            elif pos == player_pos:
-                                new_row.append("@")
-                            elif pos in box_comb:
-                                new_row.append("$")
-                            else:
-                                new_row.append(" ")
-                        level_state.append(tuple(new_row))
-
-                    state_space.add(tuple(level_state))
-        self.state_space = state_space
-        print("Finished generating state space.")
-
-    def get_action(self, key):
-        key_mapping = {
-            "w": (0, -1),
-            "a": (-1, 0),
-            "s": (0, 1),
-            "d": (1, 0),
-            "up": (0, -1),
-            "left": (-1, 0),
-            "down": (0, 1),
-            "right": (1, 0),
-        }
-        return key_mapping.get(key.lower(), None)
-
-    def execute_action(self, state, action):
-        dx, dy = action
-        x, y = self.find_player_in_state(state)
-        nx, ny = x + dx, y + dy
-        
-        reward = MALUS
-        new_position = (x, y)
-        moved_box = False
-
-        if self.is_wall(state, nx, ny):
-            return new_position, reward, moved_box
-
-        if self.is_box(state, nx, ny):
-            if not self.move_box(state, nx, ny, dx, dy):
-                return new_position, reward, moved_box
-            moved_box = True
-        
-        self.move_agent(state, x, y, nx, ny)
-        new_position = (nx, ny)
-        reward = STEPREWARD
-
-        if moved_box:
-            bx, by = nx + dx, ny + dy
-            if self.is_box_stuck(state, bx, by):
-                reward = SUPERMALUS
-            elif self.check_win(state):
-                reward = SUPERBONUS
-            elif self.is_box_placed(state, bx, by):
-                reward = BONUS
-
-        return new_position, reward, moved_box
-
-    def move_agent(self, state, x, y, nx, ny):
-        # Update player's current position
-        if state[y][x] == "@":
-            state[y][x] = " "
-        elif state[y][x] == "+":
-            state[y][x] = "."
-
-        # Update new player's position
-        if state[ny][nx] == ".":
-            state[ny][nx] = "+"
-        else:
-            state[ny][nx] = "@"
-
-    def move_box(self, state, x, y, dx, dy):
-        nx, ny = x + dx, y + dy
-        if state[ny][nx] in [" ", "."]:
-            # Update box's current position
-            if state[y][x] == "$":
-                state[y][x] = " "
-            elif state[y][x] == "x":
-                state[y][x] = "."
-
-            # Update box's new position
-            if state[ny][nx] == ".":
-                state[ny][nx] = "x"
-            else:
-                state[ny][nx] = "$"
-            return True
-        return False
-
-    def is_wall(self, state, x, y):
-        return state[y][x] == "#"
-
-    def is_box(self, state, x, y):
-        return state[y][x] in ["$", "x"]
-
-    def is_goal(self, state, x, y):
-        return state[y][x] == "."
-
-    def is_box_placed(self, state, x, y):
-        return state[y][x] in ["x"]
-
-    def is_position_free(self, state, x, y):
-        return state[y][x] in [' ', '.']
-
-    def is_box_stuck(self, state, x, y):
-        # If the box is on a goal, it's not stuck
-        if self.is_box_placed(state, x, y):
-            return False
-
-        dirs = [(0, -1), (1, 0), (0, 1), (-1, 0)]
-
-        # Check for corner configurations
-        for i in range(4):
-            dx1, dy1 = dirs[i]
-            dx2, dy2 = dirs[(i + 1) % 4]
-            if self.is_obstacle(state, x + dx1, y + dy1) and self.is_obstacle(state, x + dx2, y + dy2):
-                return True
-
-        return False
-
-    def is_obstacle(self, state, x, y):
-        return state[y][x] in ["#", "$", "x"]
-
-    def check_win(self, state):
-        for row in state:
-            if '$' in row:
-                return False
-        return True
-    
-    def serialize_state(self, state):
-        return tuple(tuple(row) for row in state)
-
-    def mc_policy_evaluation(self, num_episodes=100000, gamma=0.95, epsilon=0.9, every_visit=True, convergence_thres=0.001):        
-        print("Running Monte Carlo policy optimization algorithm...")
-        start_time = time.time()
-        Q = {}
-        returns_sum = {}
-        returns_count = {}
-        policy = {}
-        
-        for episode in range(num_episodes):
-            # print("Episode", episode + 1)
-            Q_old = copy.deepcopy(Q)
-            current_state = copy.deepcopy(self.level)
-            steps = 0
-            trajectory = []
-            terminalState = False
-
-            moved_box = False
-            prev_pos = None
-
-            while not terminalState and steps < MAXSTEPS:
-                steps += 1
-                serialized_current_state = self.serialize_state(current_state)
-
-                epsilon = max(0.1, epsilon * EPSILONDECAY)
-                
-                if random.random() < epsilon:
-                    action = random.choice(self.action_space)
-                else:
-                    action = policy.get(serialized_current_state, random.choice(self.action_space))
-                
-                action_vector = self.get_action(action)
-
-                temp_pos = self.find_player_in_state(current_state)
-                previously_moved_box = moved_box
-
-                new_pos, reward, moved_box = self.execute_action(current_state, action_vector)
-                
-                if prev_pos == new_pos and not previously_moved_box:
-                    # print("LOOP DETECTED!")
-                    terminalState = True
-                    trajectory.append((serialized_current_state, action, SUPERMALUS))
-                else:
-                    trajectory.append((serialized_current_state, action, reward))
-                
-                prev_pos = temp_pos
-
-                if reward == SUPERBONUS:
-                    # print("PUZZLE COMPLETED!")
-                    terminalState = True
-                elif reward == SUPERMALUS:
-                    # print("BOX STUCK!")
-                    terminalState = True
-
-            # print("Number of steps: " + str(steps))
-            
-            episode_length = len(trajectory)
-            visited_state_actions = set()
-
-            for t in range(episode_length):
-                state, action, reward = trajectory[t]
-                if every_visit or (state, action) not in visited_state_actions:
-                    visited_state_actions.add((state, action))
-
-                    if state not in Q:
-                        Q[state] = {a: 0.0 for a in self.action_space}
-                        returns_sum[state] = {a: 0.0 for a in self.action_space}
-                        returns_count[state] = {a: 0 for a in self.action_space}
-                    
-                    G = 0
-                    for k in range(t, episode_length):
-                        state_k, action_k, reward_k = trajectory[k]
-                        G += gamma ** (k - t) * reward_k
-                    
-                    returns_sum[state][action] += G
-                    returns_count[state][action] += 1
-                    Q[state][action] = returns_sum[state][action] / returns_count[state][action]
-
-                    best_action = max(Q[state], key=Q[state].get)
-                    policy[state] = best_action
-
-            if convergence_thres > 0 and self.has_converged(Q_old, Q, convergence_thres):
-                print("Q has converged.")
-                break
-    
-        self.policy = policy
-        self.time_to_train = time.time() - start_time
-        print("Total number of episodes: " + str(episode + 1))
-        print("Monte Carlo policy optimization completed.")
-
-    def has_converged(self, Q_old, Q_new, threshold):
-        if len(Q_old) < 5:
-            return False
-        max_diff = 0 
-        for state in Q_old:
-            if state not in Q_new:
-                continue 
-            action_old = max(Q_old[state], key=Q_old[state].get)
-            action_new = max(Q_new[state], key=Q_new[state].get)
-            if action_old != action_new:
-                return False
-            for action in Q_old[state]:
-                if action not in Q_new[state]:
-                    continue
-                diff = abs(Q_old[state][action] - Q_new[state][action])
-                max_diff = max(max_diff, diff) 
-        return max_diff < threshold
 
     def auto_play(self):
-        """
-        Automatically plays the game according to the learned policy.
-        """
         print("Playing level according to the learned policy...")
         self.number_of_actions = 0
         while not self.game_over:
-            state = self.serialize_state(self.level)
+            state = util.serialize_state(self.level)
             action = self.policy.get(state, None)
             if action is None:
                 print("No action found in policy for the current state.")
                 break
-            action_vector = self.get_action(action)
+            action_vector = util.get_action(action)
             if action_vector is None:
                 print(f"Invalid action '{action}' in policy for state.")
                 break
-            time.sleep(1)  # Wait between actions for visualization
-            self.player_pos, reward, moved_box = self.execute_action(self.level, action_vector)
+            time.sleep(1)  
+            self.player_pos, reward, moved_box = util.execute_action(self.level, action_vector)
             self.number_of_actions += 1
             self.draw_game()
             if reward == SUPERBONUS:
@@ -445,10 +137,5 @@ class Sokoban:
         print("Finished playing level.")
 
     def print_metrics(self):
-        print("Game won: " + str(self.check_win(self.level)))
-        print("Percentage of boxes placed: " + str(self.count_placed_boxes(self.level) / self.total_boxes * 100))
+        print("Percentage of boxes placed: " + str(util.count_placed_boxes(self.level) / self.total_boxes * 100))
         print("Number of actions: " + str(self.number_of_actions))
-        print(f"Time to train: : {self.time_to_train:.2f}s")
-
-    def count_placed_boxes(self, state):
-        return sum(row.count("x") for row in state)
