+++ constants.py
@@ -1,0 +1,18 @@
+SUPERMALUS = -100
+MALUS = -1
+STEPREWARD = 0
+BONUS = 1
+SUPERBONUS = 100
+MAXSTEPS = 100000
+EPSILONDECAY = 0.995
+ACTIONSPACE = ['up', 'left', 'down', 'right']
+KEYMAPPING = {
+    "w": (0, -1),
+    "a": (-1, 0),
+    "s": (0, 1),
+    "d": (1, 0),
+    "up": (0, -1),
+    "left": (-1, 0),
+    "down": (0, 1),
+    "right": (1, 0),
+}
