+++ montecarlo.py
@@ -1,6 +1,7 @@
 import sys
 import tkinter as tk
 from SokobanGame import Sokoban
+from algorithms import mc_policy_evaluation
 
 if __name__ == "__main__":
     if len(sys.argv) != 5:
@@ -10,7 +11,7 @@ if __name__ == "__main__":
     root = tk.Tk()
     game = Sokoban(master=root, level_file=sys.argv[1])
 
-    game.mc_policy_evaluation(num_episodes=int(sys.argv[2]), gamma=float(sys.argv[3]), epsilon=float(sys.argv[4]))
+    game.policy = mc_policy_evaluation(initial_state=game.level, num_episodes=int(sys.argv[2]), gamma=float(sys.argv[3]), epsilon=float(sys.argv[4]))
     game.auto_play()
     game.print_metrics()
